+++ src/app/usecases/join/join_input.go
@@ -1,0 +1,10 @@
+// Package join ゲーム参加に関するパッケージ
+package join
+
+// JoinInput DTO
+type JoinInput struct {
+	ReplyToken string
+	MemberID   string
+	MemberName string
+	GloupID    string
+}
