+++ src/app/usecases/join/join_usecase.go
@@ -1,0 +1,8 @@
+// Package join ゲーム参加に関するパッケージ
+package join
+
+// JoinUseCase ゲーム参加ユースケース
+type JoinUseCase interface {
+	// ゲームに参加する
+	JoinGame(JoinInput) JoinOutput
+}
