+++ src/app/usecases/join/join_usecase_impl.go
@@ -1,4 +1,5 @@
-package join_usecase
+// Package join ゲーム参加に関するパッケージ
+package join
 
 import (
 	"github.com/takuyaaaaaaahaaaaaa/go_wordwolf/src/app/domain"
@@ -9,10 +10,10 @@ type JoinUseCaseImpl struct {
 }
 
 // ゲームに参加する
-func (u *JoinUseCaseImpl) joinGame(JoinInputDto) JoinOutputDto {
-	// ゲームマスター排他制御
+func (u *JoinUseCaseImpl) joinGame(JoinInput) JoinOutput {
 	// ゲームマスターを取得する
 	// ゲームマスターにメンバーを追加する
-	// ゲームマスターの状態を保存する。
-	// プレゼンターに保存する
+	// メンバーを保存する
+	// ゲームマスタを保存する。
+	// プレゼンターに出力する
 }
