--- src/app/usecases/join_usecase.go
@@ -1,6 +1,0 @@
-package usecases
-
-type IJoinUseCase interface {
-	// ゲームに参加する
-	JoinGame(id int, groupId int) error
-}
