--- src/app/usecases/join_usecase/join_output_dto.go
@@ -1,8 +1,0 @@
-package join_usecase
-
-type JoinOutputDto struct {
-	ReplyToken string
-	MemberID   string
-	MemberName string
-	GloupID    string
-}
