--- src/app/usecases/join_usecase/join_usecase.go
@@ -1,6 +1,0 @@
-package join_usecase
-
-type JoinUseCase interface {
-	// ゲームに参加する
-	JoinGame(JoinInputDto) JoinOutputDto
-}
